# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A double-entry ledger and transfer orchestration service for fintech infrastructure.",
        "title": "Ledgr — Payments Engine",
        "contact": {
            "name": "Ledgr"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/accounts": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Create an account",
                "parameters": [
                    {
                        "description": "Account details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.AccountResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/accounts/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Get account by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AccountResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/accounts/{id}/audit": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Account audit log",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "2024-01-01T00:00:00Z",
                        "description": "RFC3339 start time (inclusive)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2024-12-31T23:59:59Z",
                        "description": "RFC3339 end time (inclusive)",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.AuditEventResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/accounts/{id}/entries": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "List journal entries",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.AccountEntryResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/ledger/trial-balance": {
            "get": {
                "description": "Returns debits, credits, and net for every account. The balanced field is true when net_total is zero, confirming the double-entry invariant holds.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ledger"
                ],
                "summary": "Trial balance",
                "parameters": [
                    {
                        "type": "string",
                        "example": "USD",
                        "description": "ISO 4217 currency code",
                        "name": "currency",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.TrialBalanceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.AccountEntryResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string",
                    "example": "250.0000"
                },
                "entry_description": {
                    "type": "string",
                    "example": "Transfer settlement"
                },
                "entry_id": {
                    "type": "string",
                    "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                },
                "entry_status": {
                    "type": "string",
                    "enum": [
                        "POSTED",
                        "REVERSED"
                    ],
                    "example": "POSTED"
                },
                "line_description": {
                    "type": "string",
                    "example": "Debit leg of transfer"
                },
                "line_id": {
                    "type": "string",
                    "example": "c3d4e5f6-a7b8-9012-cdef-123456789012"
                },
                "line_type": {
                    "type": "string",
                    "enum": [
                        "DEBIT",
                        "CREDIT"
                    ],
                    "example": "DEBIT"
                },
                "posted_at": {
                    "type": "string",
                    "example": "2024-01-15T10:30:00Z"
                },
                "reference": {
                    "type": "string",
                    "example": "b2c3d4e5-f6a7-8901-bcde-f12345678901"
                },
                "sequence": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "dto.AccountResponse": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "string",
                    "example": "1000.0000"
                },
                "currency": {
                    "type": "string",
                    "example": "USD"
                },
                "id": {
                    "type": "string",
                    "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                },
                "name": {
                    "type": "string",
                    "example": "Operating Account"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "ACTIVE",
                        "FROZEN",
                        "CLOSED"
                    ],
                    "example": "ACTIVE"
                },
                "type": {
                    "type": "string",
                    "example": "ASSET"
                },
                "version": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "dto.AuditEventResponse": {
            "type": "object",
            "properties": {
                "actor": {
                    "type": "string",
                    "example": "payments-engine"
                },
                "entity_id": {
                    "type": "string",
                    "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                },
                "entity_type": {
                    "type": "string",
                    "enum": [
                        "account",
                        "transaction"
                    ],
                    "example": "account"
                },
                "event_type": {
                    "type": "string",
                    "example": "account.created"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "occurred_at": {
                    "type": "string",
                    "example": "2024-01-15T10:30:00.000000000Z"
                },
                "payload": {
                    "type": "object"
                }
            }
        },
...
        "dto.CreateAccountRequest": {
            "type": "object",
            "properties": {
                "currency": {
                    "type": "string",
                    "example": "USD"
                },
                "name": {
                    "type": "string",
                    "example": "Operating Account"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "ASSET",
                        "LIABILITY",
                        "EQUITY",
                        "INCOME",
                        "EXPENSE"
                    ],
                    "example": "ASSET"
                }
            }
        },
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "account not found"
                }
            }
        },
...
        "dto.TrialBalanceResponse": {
            "type": "object",
            "properties": {
                "balanced": {
                    "type": "boolean",
                    "example": true
                },
                "currency": {
                    "type": "string",
                    "example": "USD"
                },
                "net_total": {
                    "type": "string",
                    "example": "0.0000"
                },
                "rows": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.TrialBalanceRow"
                    }
                }
            }
        },
        "dto.TrialBalanceRow": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "string",
                    "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                },
                "code": {
                    "type": "string",
                    "example": "GL_ASSET_CASH"
                },
                "name": {
                    "type": "string",
                    "example": "Cash"
                },
                "net": {
                    "type": "string",
                    "example": "2000.0000"
                },
                "total_credits": {
                    "type": "string",
                    "example": "3000.0000"
                },
                "total_debits": {
                    "type": "string",
                    "example": "5000.0000"
                }
            }
        }
    }
}